LayoutTests/editing/selection/ios/cannot-select-after-disabling-text-interaction-expected.txt

This test verifies that text selection fails when text interaction is disabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getSelection().toString() is ""
PASS successfullyParsed is true

TEST COMPLETE
Testing
